requirements.txt
diffusers==0.29.1
easydict==1.13
einops==0.8.0
matplotlib==3.8.4
mediapy==1.2.2
numpy==1.26.4
Pillow==10.3.0
torch==2.1.0
torchvision==0.16.0
xformers==0.0.22.post7
tqdm==4.66.2
accelerate==0.28.0
transformers==4.43.3
opencv-python
